feat: migrate Use to new history model

+++ core/tests/test_utils.py
@@ -63,22 +63,22 @@ class UtilsTestCase(TestCase):
 
     def test_is_admin_rights(self):
         from core.models import User, RoleRight, UserRole, Role
-        from core.utils import to_list_permissions, filter_validity
-        role = Role.objects.filter(is_system=64, *filter_validity()).first()
-        user = User.objects.filter(username="Admin", *filter_validity()).first()
+        from core.utils import to_list_permissions
+        role = Role.objects.filter(is_system=64, *Role.filter_validity()).first()
+        user = User.objects.filter(username="Admin", *User.filter_validity()).first()
         if not user:
             user = create_test_interactive_user(username="Admin", roles=[role.id])
         # removing all role but admin
         UserRole.objects.filter(
-            ~Q(role__is_system=64), user=user._u, *filter_validity()
+            ~Q(role__is_system=64), user=user._u, *UserRole.filter_validity()
         ).delete()
         # removing all admin rights
-        RoleRight.objects.filter(role__is_system=64, *filter_validity()).delete()
+        RoleRight.objects.filter(role__is_system=64, *RoleRight.filter_validity()).delete()
         rights = list(user.rights)
         rights_db = [
             rr.right_id
             for rr in RoleRight.filter_queryset()
-            .filter(role__is_system=64, *filter_validity(prefix="role__"))
+            .filter(role__is_system=64, *RoleRight.filter_validity(prefix="role__"))
             .distinct()
         ]
         self.assertEquals(len(rights_db), 0, "all roleright are not removed")
